LayoutTests/transitions/transition-end-event-multiple-01.html

<html>
<head>
  <title>Transition End Events</title>
  <style type="text/css" media="screen">
    .box {
      position: relative;
      left: 0;
      height: 100px;
      width: 100px;
      margin: 10px;
      background-color: blue;
      -webkit-transition-property: width, left, background-color, height, top;
      -webkit-transition-duration: 0.5s;
    }
  </style>
  <script src="end-event-helpers.js" type="text/javascript" charset="utf-8"></script>
  <script type="text/javascript" charset="utf-8">
    if (window.layoutTestController) {
        layoutTestController.dumpAsText();
        layoutTestController.waitUntilDone();
    }
    
    var expected = [
      ["background-color", "box1", "webkitTransitionEnd", 0.5],
      ["height", "box1", "webkitTransitionEnd", 0.5],
      ["left", "box1", "webkitTransitionEnd", 0.5],
      ["width", "box1", "webkitTransitionEnd", 0.5],
    ];
    
    function startTransition()
    {
      var boxes = document.body.getElementsByClassName('box');
      for (var i = 0; i < boxes.length; ++i) {
        boxes[i].addEventListener("webkitTransitionEnd", recordEvent, false);
        boxes[i].style.width = "150px";
        boxes[i].style.left = '200px';
        boxes[i].style.backgroundColor = 'red';
        boxes[i].style.height = '80px';
      }
      
      window.setTimeout(cleanup, 700);
    }
    
    window.addEventListener('load', startTransition, false);
  </script>
</head>
<body>

<p>Initiating a 500ms transition on the left, background-color, width and height properties of box1.</p>

<div id="container">
  <div id="box1" class="box"></div>
</div>

<div id="result"></div>

</body>
</html>
